refactor: moving visuals to UI folder for better organization
