+++ intuitivecare/src/main/java/com/br/intuitivecare/utils/ui/ProgressDialog.java
@@ -1,4 +1,4 @@
-package com.br.intuitivecare.utils;
+package com.br.intuitivecare.utils.ui;
 
 import java.awt.BorderLayout;
 import java.awt.Dimension;
